Move lottery winners update
