+++ divi/src/BlockIncentivesPopulator.h
@@ -13,6 +13,9 @@ class I_SuperblockHeightValidator;
 class I_BlockSubsidyProvider;
 class SuperblockSubsidyContainer;
 class CTransaction;
+class LotteryCoinstakeData;
+class CBlock;
+
 class BlockIncentivesPopulator
 {
 private:
@@ -40,5 +43,6 @@ public:
 
 bool IsBlockValueValid(const I_SuperblockHeightValidator& heightValidator, const CBlockRewards &nExpectedValue, CAmount nMinted, int nHeight);
 bool HasValidSuperblockPayees(const CChainParams& chainParameters, const SuperblockSubsidyContainer& superblockSubsidies, const CTransaction &txNew, const CBlockIndex* pindex);
+void CalculateLotteryWinners(const CBlock& block, const CBlockIndex *prevBlockIndex, int nHeight, LotteryCoinstakeData& coinstakeDataToUpdate);
 
 #endif // BLOCK_INCENTIVES_POPULATOR_H
